api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic==2.7.4  # Updated to resolve conflict
requests==2.31.0
tenacity==8.2.3
cachetools==5.3.2
supabase==2.0.0
langchain==0.3.10
langchain-core==0.3.62  # Updated to resolve conflict
# langgraph==0.2.60 # Commented out due to httpx conflict
langchain-community==0.3.10
langchain-google-genai==2.1.5
httpx==0.24.1  # Pinning to satisfy supabase, may break langgraph
newsapi-python==0.2.7
google-search-results==2.4.2
fmpsdk==20250102.0
alpha-vantage==2.3.1
matplotlib==3.8.2
seaborn==0.13.0
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2
# faiss-cpu==1.7.4 # Commented out due to build issues (missing headers for swig)
# sentence-transformers==2.2.2 # Commented out as it depends on faiss-cpu
aiohttp==3.9.1
pymongo==4.6.1
pdfplumber==0.10.3
python-docx==1.1.0
openpyxl==3.1.2